chore: move `tag`-related code to `tags.go`
